requirements.txt
ultralytics
opencv-python
scikit-learn
pyyaml